refactor: separate login form into client component
- Move login form to dedicated client component
- Wrap useSearchParams in Suspense boundary
- Add proper error handling and form validation
- Improve component organization and maintainability
